requirements.txt
#requirements.txt
flask
requests
python-dotenv